requirements.txt
akismet==24.5.1
amqp==5.2.0
anyio==4.4.0
appscript==1.2.5
asgiref==3.7.2
async-timeout==4.0.3
beautifulsoup4==4.12.3
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.3
django-appconf==1.0.6
django-bootstrap4-form==4.0.1
django-cleanup==8.1.0
django-cors-headers==4.3.1
django-countries==7.6.1
django-debug-toolbar==4.4.2
django-defender==0.9.8
django-fast-ratelimit==9.0.1
django-helpdesk==1.2.0
django-log-viewer==1.1.7
django-model-utils==4.5.1
django-phonenumber-field==8.0.0
django-phonenumbers==1.0.1
django-redis==5.4.0
django-reversion==5.0.12
django-rosetta==0.10.0
django-simple-history==3.7.0
django-user-accounts==3.3.2
django_csp==3.8
djangoconvertvdoctopdf==1.0.1
djangorestframework==3.14.0
docx2pdf==0.1.8
drf-yasg==1.21.7
email-reply-parser==0.5.12
et-xmlfile==1.1.0
exceptiongroup==1.2.1
futures==3.0.5
goslate==1.5.4
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.7.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
inflection==0.5.1
kombu==5.3.7
lxml==5.2.2
Markdown==3.6
mysqlclient==2.2.4
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.1
phonenumbers==8.13.40
pillow==10.3.0
pinax-invitations==8.0.0
pinax-teams==3.0.0
polib==1.2.0
prompt_toolkit==3.0.47
PyMySQL==1.1.0
pyotp==2.9.0
pypandoc==1.13
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
requests==2.32.2
requests-oauthlib==2.0.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.4.4
text-unidecode==1.3
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
